Move json database file and function
